third_party/WebKit/LayoutTests/fast/borders/border-radius-hittest-expected.txt

This test checks that div block should not get events on clicking outside the rounded border but within the bounding box of the block.

PASS successfullyParsed is true

TEST COMPLETE
PASS document.elementFromPoint(x + 5, y + 5).id is 'container'
PASS document.elementFromPoint(x + 195, y + 5).id is 'container'
PASS document.elementFromPoint(x + 5, y + 195).id is 'container'
PASS document.elementFromPoint(x + 195, y + 195).id is 'container'
PASS document.elementFromPoint(x + 100, y + 100).id is 'roundedBox'
